rockchip: Fix port/LED order on NanoPi R6S
This commit fixes the ethernet device and LED setup on the FriendlyElec
NanoPi R6S to match the physical layout/labels on the case of the device.

The actual case-label, ethernet device, and underlying NIC device,
from left-to-right on back of case:
 * "LAN2" - eth0 - rk_gmac-dwmac fe1c0000.ethernet
 * "LAN1" - eth2 - r8169 0004:41:00.0
 * "WAN"  - eth1 - r8169 0003:31:00.0

The actual case-label and LED sys-name, from left-to-right on the
front of case:
 * "WAN" - "green:wan"
 * "1"   - "green:lan"
 * "2"   - "green:lan1"

(This is the layout on my R6S box; I assume it is the same for all boxes,
but I have no idea how "ethN" device names are assigned to the underlying
NIC hardware.)

I don't think any change to the smp-affinity is required, since this
commit swaps the use of the eth1 and eth2 devices, which were already
both assigned to the "big" cores (as I understand it).

Signed-off-by: Matt Marjanović <maddog@mir.com>